Extract committee API routes to dedicated module
Move all working group/committee API routes from http.ts to a new
server/src/routes/committees.ts module:
- Admin routes (/api/admin/working-groups)
- Public routes (/api/working-groups)
- User routes (/api/me/working-groups)
- Leader routes (/api/working-groups/:slug/manage/*)

This reduces http.ts by ~1500 lines and improves code organization
by keeping committee-related routes in a dedicated module.
