+++ server/src/http.ts
@@ -25,13 +25,12 @@ import Stripe from "stripe";
 import { OrganizationDatabase, CompanyType, RevenueTier } from "./db/organization-db.js";
 import { MemberDatabase } from "./db/member-db.js";
 import { JoinRequestDatabase } from "./db/join-request-db.js";
-import { WorkingGroupDatabase } from "./db/working-group-db.js";
 import { SlackDatabase } from "./db/slack-db.js";
-import { syncSlackUsers, getSyncStatus, syncWorkingGroupMembersFromSlack, syncAllWorkingGroupMembersFromSlack } from "./slack/sync.js";
+import { syncSlackUsers, getSyncStatus } from "./slack/sync.js";
 import { isSlackConfigured, testSlackConnection } from "./slack/client.js";
 import { handleSlashCommand } from "./slack/commands.js";
 import { getCompanyDomain } from "./utils/email-domain.js";
-import { requireAuth, requireAdmin, optionalAuth, invalidateSessionCache, createRequireWorkingGroupLeader, isDevModeEnabled, getDevUser, getAvailableDevUsers, getDevSessionCookieName, DEV_USERS, type DevUserConfig } from "./middleware/auth.js";
+import { requireAuth, requireAdmin, optionalAuth, invalidateSessionCache, isDevModeEnabled, getDevUser, getAvailableDevUsers, getDevSessionCookieName, DEV_USERS, type DevUserConfig } from "./middleware/auth.js";
 import { invitationRateLimiter, orgCreationRateLimiter } from "./middleware/rate-limit.js";
 import { validateOrganizationName, validateEmail } from "./middleware/validation.js";
 import jwt from "jsonwebtoken";
@@ -41,7 +40,6 @@ import {
   notifyPaymentSucceeded,
   notifyPaymentFailed,
   notifySubscriptionCancelled,
-  notifyWorkingGroupPost,
 } from "./notifications/slack.js";
 import { createAdminRouter } from "./routes/admin.js";
 import { createAdminInsightsRouter } from "./routes/admin-insights.js";
@@ -57,6 +55,7 @@ import { processAlerts, sendDailyDigest } from "./addie/services/industry-alerts
 import { createBillingRouter } from "./routes/billing.js";
 import { createPublicBillingRouter } from "./routes/billing-public.js";
 import { createOrganizationsRouter } from "./routes/organizations.js";
+import { createCommitteeRouters } from "./routes/committees.js";
 import { sendWelcomeEmail, sendUserSignupEmail, emailDb } from "./notifications/email.js";
 import { emailPrefsDb } from "./db/email-preferences-db.js";
 import { queuePerspectiveLink, processPendingResources, processRssPerspectives } from "./addie/services/content-curator.js";
@@ -3821,387 +3820,13 @@ export class HTTPServer {
     });
 
     // ========================================
-    // Admin Working Groups API Routes
+    // Committee Routes (Working Groups, Councils, Chapters)
     // ========================================
 
-    const workingGroupDb = new WorkingGroupDatabase();
-
-    // GET /api/admin/working-groups - List all working groups
-    // Query params:
-    //   type: committee type filter (working_group, council, chapter, governance)
-    this.app.get('/api/admin/working-groups', requireAuth, requireAdmin, async (req, res) => {
-      try {
-        // Parse committee type filter (admin can see all including governance)
-        const typeParam = req.query.type as string | undefined;
-        let committeeType: 'working_group' | 'council' | 'chapter' | 'governance' | undefined;
-        if (typeParam && ['working_group', 'council', 'chapter', 'governance'].includes(typeParam)) {
-          committeeType = typeParam as typeof committeeType;
-        }
-
-        const groups = await workingGroupDb.listWorkingGroups({
-          includePrivate: true,
-          committee_type: committeeType,
-          // Admin can see governance committees
-        });
-        res.json(groups);
-      } catch (error) {
-        logger.error({ err: error }, 'List working groups error:');
-        res.status(500).json({
-          error: 'Failed to list working groups',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // GET /api/admin/working-groups/search-users - Search users for leadership selection
-    // IMPORTANT: This route must come BEFORE parameterized routes like /:id
-    //
-    // Performance strategy:
-    // 1. Query local organization_memberships table (synced from WorkOS via webhooks)
-    // 2. If table is empty, show helpful message to run backfill
-    // This is instant - no WorkOS API calls needed
-    this.app.get('/api/admin/working-groups/search-users', requireAuth, requireAdmin, async (req, res) => {
-      try {
-        const { q } = req.query;
-        if (!q || typeof q !== 'string' || q.length < 2) {
-          return res.json([]);
-        }
-
-        // Use the existing WorkingGroupDatabase method which queries local DB
-        const results = await workingGroupDb.searchUsersForLeadership(q, 20);
-        res.json(results);
-      } catch (error) {
-        // Check if it's a "table doesn't exist" error
-        const errorMessage = error instanceof Error ? error.message : 'Unknown error';
-        if (errorMessage.includes('organization_memberships') && errorMessage.includes('does not exist')) {
-          logger.warn('organization_memberships table not found - run migrations and backfill');
-          return res.status(503).json({
-            error: 'User search not yet configured',
-            message: 'Run database migrations and then call POST /api/admin/users/sync-workos to populate user data',
-          });
-        }
-
-        logger.error({ err: error }, 'Search users error:');
-        res.status(500).json({
-          error: 'Failed to search users',
-          message: errorMessage,
-        });
-      }
-    });
-
-    // GET /api/admin/working-groups/:id - Get single working group with details
-    this.app.get('/api/admin/working-groups/:id', requireAuth, requireAdmin, async (req, res) => {
-      try {
-        const { id } = req.params;
-        const group = await workingGroupDb.getWorkingGroupWithDetails(id);
-
-        if (!group) {
-          return res.status(404).json({
-            error: 'Working group not found',
-            message: `No working group found with id ${id}`
-          });
-        }
-
-        res.json(group);
-      } catch (error) {
-        logger.error({ err: error }, 'Get working group error:');
-        res.status(500).json({
-          error: 'Failed to get working group',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // POST /api/admin/working-groups - Create working group
-    this.app.post('/api/admin/working-groups', requireAuth, requireAdmin, async (req, res) => {
-      try {
-        const { name, slug, description, slack_channel_url, is_private, status, display_order,
-                leader_user_ids, committee_type, region } = req.body;
-
-        if (!name || !slug) {
-          return res.status(400).json({
-            error: 'Missing required fields',
-            message: 'Name and slug are required'
-          });
-        }
-
-        // Validate slug format
-        const slugPattern = /^[a-z0-9-]+$/;
-        if (!slugPattern.test(slug)) {
-          return res.status(400).json({
-            error: 'Invalid slug',
-            message: 'Slug must contain only lowercase letters, numbers, and hyphens'
-          });
-        }
-
-        // Validate committee_type if provided
-        if (committee_type && !['working_group', 'council', 'chapter', 'governance'].includes(committee_type)) {
-          return res.status(400).json({
-            error: 'Invalid committee type',
-            message: 'Committee type must be working_group, council, chapter, or governance'
-          });
-        }
-
-        // Region is only valid for chapters - clear it for other types
-        const finalRegion = committee_type === 'chapter' ? region : null;
-
-        // Check slug availability
-        const slugAvailable = await workingGroupDb.isSlugAvailable(slug);
-        if (!slugAvailable) {
-          return res.status(409).json({
-            error: 'Slug already exists',
-            message: `A working group with slug '${slug}' already exists`
-          });
-        }
-
-        const group = await workingGroupDb.createWorkingGroup({
-          name, slug, description, slack_channel_url, is_private, status, display_order,
-          leader_user_ids, committee_type, region: finalRegion
-        });
-
-        res.status(201).json(group);
-      } catch (error) {
-        logger.error({ err: error }, 'Create working group error:');
-        res.status(500).json({
-          error: 'Failed to create working group',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // PUT /api/admin/working-groups/:id - Update working group
-    this.app.put('/api/admin/working-groups/:id', requireAuth, requireAdmin, async (req, res) => {
-      try {
-        const { id } = req.params;
-        const updates = req.body;
-
-        // Validate committee_type if provided
-        if (updates.committee_type && !['working_group', 'council', 'chapter', 'governance'].includes(updates.committee_type)) {
-          return res.status(400).json({
-            error: 'Invalid committee type',
-            message: 'Committee type must be working_group, council, chapter, or governance'
-          });
-        }
-
-        // Region is only valid for chapters - clear it for other types if committee_type is being changed
-        if (updates.committee_type && updates.committee_type !== 'chapter') {
-          updates.region = null;
-        }
-
-        const group = await workingGroupDb.updateWorkingGroup(id, updates);
-
-        if (!group) {
-          return res.status(404).json({
-            error: 'Working group not found',
-            message: `No working group found with id ${id}`
-          });
-        }
-
-        res.json(group);
-      } catch (error) {
-        logger.error({ err: error }, 'Update working group error:');
-        res.status(500).json({
-          error: 'Failed to update working group',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // DELETE /api/admin/working-groups/:id - Delete working group
-    this.app.delete('/api/admin/working-groups/:id', requireAuth, requireAdmin, async (req, res) => {
-      try {
-        const { id } = req.params;
-        const deleted = await workingGroupDb.deleteWorkingGroup(id);
-
-        if (!deleted) {
-          return res.status(404).json({
-            error: 'Working group not found',
-            message: `No working group found with id ${id}`
-          });
-        }
-
-        res.json({ success: true, deleted: id });
-      } catch (error) {
-        logger.error({ err: error }, 'Delete working group error:');
-        res.status(500).json({
-          error: 'Failed to delete working group',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // GET /api/admin/working-groups/:id/members - List working group members
-    this.app.get('/api/admin/working-groups/:id/members', requireAuth, requireAdmin, async (req, res) => {
-      try {
-        const { id } = req.params;
-        const members = await workingGroupDb.getMembershipsByWorkingGroup(id);
-        res.json(members);
-      } catch (error) {
-        logger.error({ err: error }, 'List working group members error:');
-        res.status(500).json({
-          error: 'Failed to list members',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // POST /api/admin/working-groups/:id/members - Add member to working group
-    this.app.post('/api/admin/working-groups/:id/members', requireAuth, requireAdmin, async (req, res) => {
-      try {
-        const { id } = req.params;
-        const { workos_user_id, user_email, user_name, user_org_name, workos_organization_id } = req.body;
-        const user = req.user!;
-
-        if (!workos_user_id) {
-          return res.status(400).json({
-            error: 'Missing required field',
-            message: 'workos_user_id is required'
-          });
-        }
-
-        const membership = await workingGroupDb.addMembership({
-          working_group_id: id,
-          workos_user_id,
-          user_email,
-          user_name,
-          user_org_name,
-          workos_organization_id,
-          added_by_user_id: user.id,
-        });
-
-        // Invalidate member context cache (working_groups field changed)
-        invalidateMemberContextCache();
-
-        res.status(201).json(membership);
-      } catch (error) {
-        logger.error({ err: error }, 'Add working group member error:');
-        res.status(500).json({
-          error: 'Failed to add member',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // DELETE /api/admin/working-groups/:id/members/:userId - Remove member from working group
-    this.app.delete('/api/admin/working-groups/:id/members/:userId', requireAuth, requireAdmin, async (req, res) => {
-      try {
-        const { id, userId } = req.params;
-        const deleted = await workingGroupDb.deleteMembership(id, userId);
-
-        if (!deleted) {
-          return res.status(404).json({
-            error: 'Membership not found',
-            message: 'User is not a member of this working group'
-          });
-        }
-
-        // Invalidate member context cache (working_groups field changed)
-        invalidateMemberContextCache();
-
-        res.json({ success: true });
-      } catch (error) {
-        logger.error({ err: error }, 'Remove working group member error:');
-        res.status(500).json({
-          error: 'Failed to remove member',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // POST /api/admin/working-groups/:id/sync-from-slack - Sync members from Slack channel
-    this.app.post('/api/admin/working-groups/:id/sync-from-slack', requireAuth, requireAdmin, async (req, res) => {
-      try {
-        const { id } = req.params;
-
-        // Verify the working group exists
-        const workingGroup = await workingGroupDb.getWorkingGroupById(id);
-        if (!workingGroup) {
-          return res.status(404).json({
-            error: 'Working group not found',
-            message: 'The specified working group does not exist'
-          });
-        }
-
-        // Sync members from Slack
-        const result = await syncWorkingGroupMembersFromSlack(id);
-
-        if (result.errors.length > 0 && result.members_added === 0 && result.members_already_in_group === 0) {
-          return res.status(400).json({
-            error: 'Sync failed',
-            message: result.errors[0],
-            result
-          });
-        }
-
-        // Invalidate member context cache if any members were added
-        if (result.members_added > 0) {
-          invalidateMemberContextCache();
-        }
-
-        res.json({
-          success: true,
-          result
-        });
-      } catch (error) {
-        logger.error({ err: error }, 'Sync working group members from Slack error:');
-        res.status(500).json({
-          error: 'Failed to sync members',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // POST /api/admin/working-groups/sync-all-from-slack - Sync all working groups with Slack channels
-    this.app.post('/api/admin/working-groups/sync-all-from-slack', requireAuth, requireAdmin, async (req, res) => {
-      try {
-        const results = await syncAllWorkingGroupMembersFromSlack();
-
-        const totalAdded = results.reduce((sum, r) => sum + r.members_added, 0);
-        const totalErrors = results.reduce((sum, r) => sum + r.errors.length, 0);
-
-        res.json({
-          success: true,
-          summary: {
-            groups_synced: results.length,
-            total_members_added: totalAdded,
-            total_errors: totalErrors
-          },
-          results
-        });
-      } catch (error) {
-        logger.error({ err: error }, 'Sync all working groups from Slack error:');
-        res.status(500).json({
-          error: 'Failed to sync working groups',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // GET /api/admin/working-groups/:id/posts - List all posts for a working group
-    this.app.get('/api/admin/working-groups/:id/posts', requireAuth, requireAdmin, async (req, res) => {
-      try {
-        const { id } = req.params;
-        const pool = getPool();
-
-        const result = await pool.query(
-          `SELECT id, slug, content_type, title, subtitle, category, excerpt,
-            external_url, external_site_name, author_name, author_title,
-            author_user_id, featured_image_url, status, published_at, display_order, tags
-          FROM perspectives
-          WHERE working_group_id = $1
-          ORDER BY published_at DESC NULLS LAST, created_at DESC`,
-          [id]
-        );
-
-        res.json(result.rows);
-      } catch (error) {
-        logger.error({ err: error }, 'List working group posts error:');
-        res.status(500).json({
-          error: 'Failed to list posts',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
+    const { adminApiRouter, publicApiRouter, userApiRouter } = createCommitteeRouters();
+    this.app.use('/api/admin/working-groups', adminApiRouter);
+    this.app.use('/api/working-groups', publicApiRouter);
+    this.app.use('/api/me/working-groups', userApiRouter);
 
     // ========================================
     // SEO Routes (sitemap.xml, robots.txt)
@@ -6315,1150 +5940,6 @@ Disallow: /api/admin/
         logger.error({ err: error }, 'Get member error');
         res.status(500).json({
           error: 'Failed to get member',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // ========================================
-    // Public Working Groups API Routes
-    // ========================================
-
-    // GET /api/working-groups - List active working groups (public groups for everyone, private for members)
-    // Query params:
-    //   type: committee type filter (working_group, council, chapter, or comma-separated list)
-    this.app.get('/api/working-groups', optionalAuth, async (req, res) => {
-      try {
-        const wgDb = new WorkingGroupDatabase();
-        const user = req.user; // May be undefined for anonymous users
-
-        // Parse committee type filter
-        const typeParam = req.query.type as string | undefined;
-        let committeeTypes: ('working_group' | 'council' | 'chapter' | 'governance')[] | undefined;
-        if (typeParam) {
-          const validTypes = ['working_group', 'council', 'chapter', 'governance'] as const;
-          const filtered = typeParam.split(',').filter(t =>
-            validTypes.includes(t as typeof validTypes[number])
-          ) as typeof committeeTypes;
-          committeeTypes = filtered && filtered.length > 0 ? filtered : undefined;
-        }
-
-        let groups;
-        if (user?.id) {
-          // Authenticated user - show public + private groups they're a member of
-          groups = await wgDb.listWorkingGroupsForUser(user.id, {
-            committee_type: committeeTypes,
-            excludeGovernance: true, // Always hide governance from public listings
-          });
-        } else {
-          // Anonymous user - show only public active groups
-          groups = await wgDb.listWorkingGroups({
-            status: 'active',
-            includePrivate: false,
-            committee_type: committeeTypes,
-            excludeGovernance: true, // Always hide governance from public listings
-          });
-        }
-
-        res.json({ working_groups: groups });
-      } catch (error) {
-        logger.error({ err: error }, 'List working groups error');
-        res.status(500).json({
-          error: 'Failed to list working groups',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // GET /api/working-groups/:slug - Get working group details
-    this.app.get('/api/working-groups/:slug', optionalAuth, async (req, res) => {
-      try {
-        const { slug } = req.params;
-        const wgDb = new WorkingGroupDatabase();
-        const user = req.user; // May be undefined for anonymous users
-
-        const group = await wgDb.getWorkingGroupBySlug(slug);
-
-        if (!group || group.status !== 'active') {
-          return res.status(404).json({
-            error: 'Working group not found',
-            message: `No working group found with slug: ${slug}`,
-          });
-        }
-
-        // Check access for private groups
-        if (group.is_private) {
-          if (!user?.id) {
-            return res.status(404).json({
-              error: 'Working group not found',
-              message: `No working group found with slug: ${slug}`,
-            });
-          }
-
-          const isMember = await wgDb.isMember(group.id, user.id);
-          if (!isMember) {
-            return res.status(404).json({
-              error: 'Working group not found',
-              message: `No working group found with slug: ${slug}`,
-            });
-          }
-        }
-
-        // Get memberships for display
-        const memberships = await wgDb.getMembershipsByWorkingGroup(group.id);
-
-        // Check if current user is a member
-        let isMember = false;
-        if (user?.id) {
-          isMember = await wgDb.isMember(group.id, user.id);
-        }
-
-        res.json({
-          working_group: {
-            ...group,
-            member_count: memberships.length,
-            memberships,
-          },
-          is_member: isMember,
-        });
-      } catch (error) {
-        logger.error({ err: error }, 'Get working group error');
-        res.status(500).json({
-          error: 'Failed to get working group',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // GET /api/working-groups/:slug/posts - Get published posts for a working group
-    this.app.get('/api/working-groups/:slug/posts', optionalAuth, async (req, res) => {
-      try {
-        const { slug } = req.params;
-        const wgDb = new WorkingGroupDatabase();
-        const pool = getPool();
-        const user = req.user; // May be undefined for anonymous users
-
-        const group = await wgDb.getWorkingGroupBySlug(slug);
-
-        if (!group || group.status !== 'active') {
-          return res.status(404).json({
-            error: 'Working group not found',
-            message: `No working group found with slug: ${slug}`,
-          });
-        }
-
-        // Check access for private groups and determine membership
-        let isMember = false;
-        if (user?.id) {
-          isMember = await wgDb.isMember(group.id, user.id);
-        }
-
-        if (group.is_private) {
-          if (!user?.id || !isMember) {
-            return res.status(404).json({
-              error: 'Working group not found',
-              message: `No working group found with slug: ${slug}`,
-            });
-          }
-        }
-
-        // If user is a member, show all posts; otherwise filter out members-only posts
-        const result = await pool.query(
-          `SELECT id, slug, content_type, title, subtitle, category, excerpt,
-            external_url, external_site_name, author_name, author_title,
-            featured_image_url, published_at, tags, is_members_only
-          FROM perspectives
-          WHERE working_group_id = $1 AND status = 'published'
-            AND (is_members_only = false OR $2 = true)
-          ORDER BY published_at DESC NULLS LAST`,
-          [group.id, isMember]
-        );
-
-        res.json({ posts: result.rows });
-      } catch (error) {
-        logger.error({ err: error }, 'Get working group posts error');
-        res.status(500).json({
-          error: 'Failed to get posts',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // POST /api/working-groups/:slug/join - Join a public working group
-    this.app.post('/api/working-groups/:slug/join', requireAuth, async (req, res) => {
-      try {
-        const { slug } = req.params;
-        const wgDb = new WorkingGroupDatabase();
-        const user = req.user!;
-
-        const group = await wgDb.getWorkingGroupBySlug(slug);
-
-        if (!group || group.status !== 'active') {
-          return res.status(404).json({
-            error: 'Working group not found',
-            message: `No working group found with slug: ${slug}`,
-          });
-        }
-
-        if (group.is_private) {
-          return res.status(403).json({
-            error: 'Private group',
-            message: 'This working group is private and requires an invitation to join',
-          });
-        }
-
-        // Check if already a member
-        const existingMembership = await wgDb.getMembership(group.id, user.id);
-        if (existingMembership && existingMembership.status === 'active') {
-          return res.status(409).json({
-            error: 'Already a member',
-            message: 'You are already a member of this working group',
-          });
-        }
-
-        // Get user's organization info for the membership record
-        let orgId: string | undefined;
-        let orgName: string | undefined;
-        if (workos) {
-          try {
-            const memberships = await workos.userManagement.listOrganizationMemberships({
-              userId: user.id,
-            });
-            if (memberships.data.length > 0) {
-              const org = await workos.organizations.getOrganization(memberships.data[0].organizationId);
-              orgId = org.id;
-              orgName = org.name;
-            }
-          } catch {
-            // Ignore org fetch errors
-          }
-        }
-
-        const membership = await wgDb.addMembership({
-          working_group_id: group.id,
-          workos_user_id: user.id,
-          user_email: user.email,
-          user_name: user.firstName && user.lastName
-            ? `${user.firstName} ${user.lastName}`
-            : user.email,
-          workos_organization_id: orgId,
-          user_org_name: orgName,
-          added_by_user_id: user.id, // Self-join
-        });
-
-        // Invalidate Addie's member context cache - working group membership changed
-        invalidateMemberContextCache();
-
-        res.status(201).json({ success: true, membership });
-      } catch (error) {
-        logger.error({ err: error }, 'Join working group error');
-        res.status(500).json({
-          error: 'Failed to join working group',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // DELETE /api/working-groups/:slug/leave - Leave a working group
-    this.app.delete('/api/working-groups/:slug/leave', requireAuth, async (req, res) => {
-      try {
-        const { slug } = req.params;
-        const wgDb = new WorkingGroupDatabase();
-        const user = req.user!;
-
-        const group = await wgDb.getWorkingGroupBySlug(slug);
-
-        if (!group) {
-          return res.status(404).json({
-            error: 'Working group not found',
-            message: `No working group found with slug: ${slug}`,
-          });
-        }
-
-        // Check if user is a leader - they can't leave without being replaced
-        const isLeader = group.leaders?.some(l => l.user_id === user.id) ?? false;
-        if (isLeader) {
-          return res.status(403).json({
-            error: 'Cannot leave',
-            message: 'As a leader, you must be replaced before leaving the group',
-          });
-        }
-
-        const removed = await wgDb.removeMembership(group.id, user.id);
-
-        if (!removed) {
-          return res.status(404).json({
-            error: 'Not a member',
-            message: 'You are not a member of this working group',
-          });
-        }
-
-        // Invalidate Addie's member context cache - working group membership changed
-        invalidateMemberContextCache();
-
-        res.json({ success: true });
-      } catch (error) {
-        logger.error({ err: error }, 'Leave working group error');
-        res.status(500).json({
-          error: 'Failed to leave working group',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // GET /api/me/working-groups - Get current user's working group memberships
-    this.app.get('/api/me/working-groups', requireAuth, async (req, res) => {
-      try {
-        const wgDb = new WorkingGroupDatabase();
-        const user = req.user!;
-
-        const groups = await wgDb.getWorkingGroupsForUser(user.id);
-        res.json({ working_groups: groups });
-      } catch (error) {
-        logger.error({ err: error }, 'Get user working groups error');
-        res.status(500).json({
-          error: 'Failed to get working groups',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // POST /api/working-groups/:slug/posts - Create a post in a working group (members)
-    // Members can only create members-only posts; leaders can create public posts
-    this.app.post('/api/working-groups/:slug/posts', requireAuth, async (req, res) => {
-      try {
-        const { slug } = req.params;
-        const { title, content, content_type, category, excerpt, external_url, external_site_name, post_slug, is_members_only } = req.body;
-        const wgDb = new WorkingGroupDatabase();
-        const pool = getPool();
-        const user = req.user!;
-
-        const group = await wgDb.getWorkingGroupBySlug(slug);
-
-        if (!group || group.status !== 'active') {
-          return res.status(404).json({
-            error: 'Working group not found',
-            message: `No working group found with slug: ${slug}`,
-          });
-        }
-
-        // Check if user is a member
-        const isMember = await wgDb.isMember(group.id, user.id);
-        if (!isMember) {
-          return res.status(403).json({
-            error: 'Not a member',
-            message: 'You must be a member of this working group to post',
-          });
-        }
-
-        // Check if user is a leader
-        const isLeader = group.leaders?.some(l => l.user_id === user.id) ?? false;
-
-        // Non-leaders can only create members-only posts
-        const finalMembersOnly = isLeader ? (is_members_only ?? true) : true;
-
-        if (!title || !post_slug) {
-          return res.status(400).json({
-            error: 'Missing required fields',
-            message: 'Title and slug are required',
-          });
-        }
-
-        // Validate slug format
-        const slugPattern = /^[a-z0-9-]+$/;
-        if (!slugPattern.test(post_slug)) {
-          return res.status(400).json({
-            error: 'Invalid slug',
-            message: 'Slug must contain only lowercase letters, numbers, and hyphens',
-          });
-        }
-
-        // Create the post (perspective with working_group_id)
-        const authorName = user.firstName && user.lastName
-          ? `${user.firstName} ${user.lastName}`
-          : user.email;
-
-        const result = await pool.query(
-          `INSERT INTO perspectives (
-            working_group_id, slug, content_type, title, content, category, excerpt,
-            external_url, external_site_name, author_name, author_user_id,
-            status, published_at, is_members_only
-          ) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, 'published', NOW(), $12)
-          RETURNING *`,
-          [
-            group.id,
-            post_slug,
-            content_type || 'article',
-            title,
-            content || null,
-            category || null,
-            excerpt || null,
-            external_url || null,
-            external_site_name || null,
-            authorName,
-            user.id,
-            finalMembersOnly,
-          ]
-        );
-
-        // Send Slack notification for public posts
-        if (!finalMembersOnly) {
-          notifyWorkingGroupPost({
-            workingGroupName: group.name,
-            workingGroupSlug: slug,
-            postTitle: title,
-            postSlug: post_slug,
-            authorName,
-            contentType: content_type || 'article',
-            category: category || undefined,
-          }).catch(err => {
-            logger.warn({ err }, 'Failed to send Slack notification for working group post');
-          });
-        }
-
-        res.status(201).json({ post: result.rows[0] });
-      } catch (error) {
-        logger.error({ err: error }, 'Create working group post error');
-        if (error instanceof Error && error.message.includes('duplicate key')) {
-          return res.status(409).json({
-            error: 'Slug already exists',
-            message: 'A post with this slug already exists in this working group',
-          });
-        }
-        res.status(500).json({
-          error: 'Failed to create post',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // PUT /api/working-groups/:slug/posts/:postId - Update own post (members)
-    this.app.put('/api/working-groups/:slug/posts/:postId', requireAuth, async (req, res) => {
-      try {
-        const { slug, postId } = req.params;
-        const { title, content, content_type, category, excerpt, external_url, external_site_name, post_slug, is_members_only } = req.body;
-        const wgDb = new WorkingGroupDatabase();
-        const pool = getPool();
-        const user = req.user!;
-
-        const group = await wgDb.getWorkingGroupBySlug(slug);
-
-        if (!group || group.status !== 'active') {
-          return res.status(404).json({
-            error: 'Working group not found',
-            message: `No working group found with slug: ${slug}`,
-          });
-        }
-
-        // Check if user is a member
-        const isMember = await wgDb.isMember(group.id, user.id);
-        if (!isMember) {
-          return res.status(403).json({
-            error: 'Not a member',
-            message: 'You must be a member of this working group',
-          });
-        }
-
-        // Get existing post
-        const existing = await pool.query(
-          `SELECT * FROM perspectives WHERE id = $1 AND working_group_id = $2`,
-          [postId, group.id]
-        );
-
-        if (existing.rows.length === 0) {
-          return res.status(404).json({
-            error: 'Post not found',
-            message: 'Post not found in this working group',
-          });
-        }
-
-        const post = existing.rows[0];
-        const isLeader = group.leaders?.some(l => l.user_id === user.id) ?? false;
-        const isAuthor = post.author_user_id === user.id;
-
-        // Only authors or leaders can edit posts
-        if (!isAuthor && !isLeader) {
-          return res.status(403).json({
-            error: 'Not authorized',
-            message: 'You can only edit your own posts',
-          });
-        }
-
-        // Non-leaders cannot make posts public
-        const finalMembersOnly = isLeader ? (is_members_only ?? post.is_members_only) : true;
-
-        // Validate slug if changing
-        if (post_slug && post_slug !== post.slug) {
-          const slugPattern = /^[a-z0-9-]+$/;
-          if (!slugPattern.test(post_slug)) {
-            return res.status(400).json({
-              error: 'Invalid slug',
-              message: 'Slug must contain only lowercase letters, numbers, and hyphens',
-            });
-          }
-        }
-
-        const result = await pool.query(
-          `UPDATE perspectives SET
-            slug = COALESCE($1, slug),
-            content_type = COALESCE($2, content_type),
-            title = COALESCE($3, title),
-            content = $4,
-            category = $5,
-            excerpt = $6,
-            external_url = $7,
-            external_site_name = $8,
-            is_members_only = $9,
-            updated_at = NOW()
-          WHERE id = $10 AND working_group_id = $11
-          RETURNING *`,
-          [
-            post_slug || null,
-            content_type || null,
-            title || null,
-            content ?? post.content,
-            category ?? post.category,
-            excerpt ?? post.excerpt,
-            external_url ?? post.external_url,
-            external_site_name ?? post.external_site_name,
-            finalMembersOnly,
-            postId,
-            group.id,
-          ]
-        );
-
-        res.json({ post: result.rows[0] });
-      } catch (error) {
-        logger.error({ err: error }, 'Update working group post error');
-        if (error instanceof Error && error.message.includes('duplicate key')) {
-          return res.status(409).json({
-            error: 'Slug already exists',
-            message: 'A post with this slug already exists',
-          });
-        }
-        res.status(500).json({
-          error: 'Failed to update post',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // DELETE /api/working-groups/:slug/posts/:postId - Delete own post (members)
-    this.app.delete('/api/working-groups/:slug/posts/:postId', requireAuth, async (req, res) => {
-      try {
-        const { slug, postId } = req.params;
-        const wgDb = new WorkingGroupDatabase();
-        const pool = getPool();
-        const user = req.user!;
-
-        const group = await wgDb.getWorkingGroupBySlug(slug);
-
-        if (!group || group.status !== 'active') {
-          return res.status(404).json({
-            error: 'Working group not found',
-            message: `No working group found with slug: ${slug}`,
-          });
-        }
-
-        // Check if user is a member
-        const isMember = await wgDb.isMember(group.id, user.id);
-        if (!isMember) {
-          return res.status(403).json({
-            error: 'Not a member',
-            message: 'You must be a member of this working group',
-          });
-        }
-
-        // Get existing post
-        const existing = await pool.query(
-          `SELECT * FROM perspectives WHERE id = $1 AND working_group_id = $2`,
-          [postId, group.id]
-        );
-
-        if (existing.rows.length === 0) {
-          return res.status(404).json({
-            error: 'Post not found',
-            message: 'Post not found in this working group',
-          });
-        }
-
-        const post = existing.rows[0];
-        const isLeader = group.leaders?.some(l => l.user_id === user.id) ?? false;
-        const isAuthor = post.author_user_id === user.id;
-
-        // Only authors or leaders can delete posts
-        if (!isAuthor && !isLeader) {
-          return res.status(403).json({
-            error: 'Not authorized',
-            message: 'You can only delete your own posts',
-          });
-        }
-
-        await pool.query(
-          `DELETE FROM perspectives WHERE id = $1 AND working_group_id = $2`,
-          [postId, group.id]
-        );
-
-        res.json({ success: true });
-      } catch (error) {
-        logger.error({ err: error }, 'Delete working group post error');
-        res.status(500).json({
-          error: 'Failed to delete post',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // POST /api/working-groups/:slug/fetch-url - Fetch URL metadata (for link posts) - members only
-    this.app.post('/api/working-groups/:slug/fetch-url', requireAuth, async (req, res) => {
-      try {
-        const { slug } = req.params;
-        const { url } = req.body;
-        const wgDb = new WorkingGroupDatabase();
-        const user = req.user!;
-
-        const group = await wgDb.getWorkingGroupBySlug(slug);
-
-        if (!group || group.status !== 'active') {
-          return res.status(404).json({
-            error: 'Working group not found',
-            message: `No working group found with slug: ${slug}`,
-          });
-        }
-
-        // Check if user is a member
-        const isMember = await wgDb.isMember(group.id, user.id);
-        if (!isMember) {
-          return res.status(403).json({
-            error: 'Not a member',
-            message: 'You must be a member of this working group',
-          });
-        }
-
-        if (!url) {
-          return res.status(400).json({
-            error: 'URL required',
-            message: 'Please provide a URL to fetch',
-          });
-        }
-
-        // Fetch the page
-        const response = await fetch(url, {
-          headers: {
-            'User-Agent': 'Mozilla/5.0 (compatible; AgenticAdvertising/1.0)',
-            'Accept': 'text/html,application/xhtml+xml',
-          },
-          redirect: 'follow',
-        });
-
-        if (!response.ok) {
-          throw new Error(`Failed to fetch URL: ${response.status}`);
-        }
-
-        const html = await response.text();
-
-        // Extract metadata from HTML
-        const titleMatch = html.match(/<title[^>]*>([^<]+)<\/title>/i);
-        const ogTitleMatch = html.match(/<meta[^>]+property=["']og:title["'][^>]+content=["']([^"']+)["']/i)
-          || html.match(/<meta[^>]+content=["']([^"']+)["'][^>]+property=["']og:title["']/i);
-        const ogDescMatch = html.match(/<meta[^>]+property=["']og:description["'][^>]+content=["']([^"']+)["']/i)
-          || html.match(/<meta[^>]+content=["']([^"']+)["'][^>]+property=["']og:description["']/i);
-        const descMatch = html.match(/<meta[^>]+name=["']description["'][^>]+content=["']([^"']+)["']/i)
-          || html.match(/<meta[^>]+content=["']([^"']+)["'][^>]+name=["']description["']/i);
-        const ogSiteMatch = html.match(/<meta[^>]+property=["']og:site_name["'][^>]+content=["']([^"']+)["']/i)
-          || html.match(/<meta[^>]+content=["']([^"']+)["'][^>]+property=["']og:site_name["']/i);
-
-        // Helper to decode HTML entities
-        const decodeHtmlEntities = (text: string): string => {
-          return text
-            .replace(/&#x([0-9a-fA-F]+);/g, (_, hex) => String.fromCharCode(parseInt(hex, 16)))
-            .replace(/&#(\d+);/g, (_, dec) => String.fromCharCode(parseInt(dec, 10)))
-            .replace(/&quot;/g, '"')
-            .replace(/&apos;/g, "'")
-            .replace(/&lt;/g, '<')
-            .replace(/&gt;/g, '>')
-            .replace(/&amp;/g, '&');
-        };
-
-        let title = ogTitleMatch?.[1] || titleMatch?.[1] || '';
-        title = decodeHtmlEntities(title.trim());
-
-        let excerpt = ogDescMatch?.[1] || descMatch?.[1] || '';
-        excerpt = decodeHtmlEntities(excerpt.trim());
-        // Truncate excerpt to 160 chars max
-        if (excerpt.length > 160) {
-          excerpt = excerpt.substring(0, 157) + '...';
-        }
-
-        let site_name = ogSiteMatch?.[1] || '';
-        if (!site_name) {
-          try {
-            const parsedUrl = new URL(url);
-            site_name = parsedUrl.hostname.replace('www.', '');
-            site_name = site_name.charAt(0).toUpperCase() + site_name.slice(1);
-          } catch {
-            // ignore URL parse errors
-          }
-        }
-        site_name = decodeHtmlEntities(site_name);
-
-        res.json({ title, excerpt, site_name });
-      } catch (error) {
-        logger.error({ err: error }, 'Fetch URL metadata error (member)');
-        res.status(500).json({
-          error: 'Failed to fetch URL',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // ========================================
-    // Working Group Leader API Routes (Chair/Vice-Chair only)
-    // ========================================
-
-    const wgDbForLeader = new WorkingGroupDatabase();
-    const requireWorkingGroupLeader = createRequireWorkingGroupLeader(wgDbForLeader);
-
-    // GET /api/working-groups/:slug/manage/posts - List all posts (including drafts) for leaders
-    this.app.get('/api/working-groups/:slug/manage/posts', requireAuth, requireWorkingGroupLeader, async (req, res) => {
-      try {
-        const { slug } = req.params;
-        const pool = getPool();
-
-        const group = await wgDbForLeader.getWorkingGroupBySlug(slug);
-        if (!group) {
-          return res.status(404).json({
-            error: 'Working group not found',
-            message: `No working group found with slug: ${slug}`,
-          });
-        }
-
-        const result = await pool.query(
-          `SELECT id, slug, content_type, title, subtitle, category, excerpt, content,
-            external_url, external_site_name, author_name, author_title,
-            author_user_id, featured_image_url, status, published_at, display_order, tags,
-            created_at, updated_at
-          FROM perspectives
-          WHERE working_group_id = $1
-          ORDER BY display_order ASC, published_at DESC NULLS LAST, created_at DESC`,
-          [group.id]
-        );
-
-        res.json({ posts: result.rows });
-      } catch (error) {
-        logger.error({ err: error }, 'List working group leader posts error');
-        res.status(500).json({
-          error: 'Failed to list posts',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // GET /api/working-groups/:slug/manage/posts/:postId - Get single post for editing
-    this.app.get('/api/working-groups/:slug/manage/posts/:postId', requireAuth, requireWorkingGroupLeader, async (req, res) => {
-      try {
-        const { slug, postId } = req.params;
-        const pool = getPool();
-
-        const group = await wgDbForLeader.getWorkingGroupBySlug(slug);
-        if (!group) {
-          return res.status(404).json({
-            error: 'Working group not found',
-            message: `No working group found with slug: ${slug}`,
-          });
-        }
-
-        const result = await pool.query(
-          `SELECT * FROM perspectives WHERE id = $1 AND working_group_id = $2`,
-          [postId, group.id]
-        );
-
-        if (result.rows.length === 0) {
-          return res.status(404).json({
-            error: 'Post not found',
-            message: 'Post not found in this working group',
-          });
-        }
-
-        res.json(result.rows[0]);
-      } catch (error) {
-        logger.error({ err: error }, 'Get working group post error');
-        res.status(500).json({
-          error: 'Failed to get post',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // POST /api/working-groups/:slug/manage/posts - Create post as leader (with draft support)
-    this.app.post('/api/working-groups/:slug/manage/posts', requireAuth, requireWorkingGroupLeader, async (req, res) => {
-      try {
-        const { slug } = req.params;
-        const {
-          post_slug, content_type, title, subtitle, category, excerpt, content,
-          external_url, external_site_name, author_name, author_title,
-          featured_image_url, status, display_order, tags, is_members_only
-        } = req.body;
-        const pool = getPool();
-        const user = req.user!;
-
-        const group = await wgDbForLeader.getWorkingGroupBySlug(slug);
-        if (!group) {
-          return res.status(404).json({
-            error: 'Working group not found',
-            message: `No working group found with slug: ${slug}`,
-          });
-        }
-
-        if (!title || !post_slug) {
-          return res.status(400).json({
-            error: 'Missing required fields',
-            message: 'Title and slug are required',
-          });
-        }
-
-        // Validate slug format
-        const slugPattern = /^[a-z0-9-]+$/;
-        if (!slugPattern.test(post_slug)) {
-          return res.status(400).json({
-            error: 'Invalid slug',
-            message: 'Slug must contain only lowercase letters, numbers, and hyphens',
-          });
-        }
-
-        // Validate content type for links
-        if (content_type === 'link' && !external_url) {
-          return res.status(400).json({
-            error: 'Missing external URL',
-            message: 'External URL is required for link type posts',
-          });
-        }
-
-        const authorNameFinal = author_name || (user.firstName && user.lastName
-          ? `${user.firstName} ${user.lastName}`
-          : user.email);
-
-        const result = await pool.query(
-          `INSERT INTO perspectives (
-            working_group_id, slug, content_type, title, subtitle, category, excerpt, content,
-            external_url, external_site_name, author_name, author_title, author_user_id,
-            featured_image_url, status, display_order, tags, published_at, is_members_only
-          ) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15, $16, $17, $18, $19)
-          RETURNING *`,
-          [
-            group.id,
-            post_slug,
-            content_type || 'article',
-            title,
-            subtitle || null,
-            category || null,
-            excerpt || null,
-            content || null,
-            external_url || null,
-            external_site_name || null,
-            authorNameFinal,
-            author_title || null,
-            user.id,
-            featured_image_url || null,
-            status || 'draft',
-            display_order || 0,
-            tags || null,
-            status === 'published' ? new Date() : null,
-            is_members_only || false,
-          ]
-        );
-
-        const createdPost = result.rows[0];
-
-        // Send Slack notification if post is published
-        if (status === 'published') {
-          notifyWorkingGroupPost({
-            workingGroupName: group.name,
-            workingGroupSlug: slug,
-            postTitle: title,
-            postSlug: post_slug,
-            authorName: authorNameFinal,
-            contentType: content_type || 'article',
-            category: category || undefined,
-          }).catch(err => {
-            logger.warn({ err }, 'Failed to send Slack notification for working group post');
-          });
-        }
-
-        res.status(201).json(createdPost);
-      } catch (error) {
-        logger.error({ err: error }, 'Create working group leader post error');
-        if (error instanceof Error && error.message.includes('duplicate key')) {
-          return res.status(409).json({
-            error: 'Slug already exists',
-            message: 'A post with this slug already exists',
-          });
-        }
-        res.status(500).json({
-          error: 'Failed to create post',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // PUT /api/working-groups/:slug/manage/posts/:postId - Update post as leader
-    this.app.put('/api/working-groups/:slug/manage/posts/:postId', requireAuth, requireWorkingGroupLeader, async (req, res) => {
-      try {
-        const { slug, postId } = req.params;
-        const {
-          post_slug, content_type, title, subtitle, category, excerpt, content,
-          external_url, external_site_name, author_name, author_title,
-          featured_image_url, status, display_order, tags, is_members_only
-        } = req.body;
-        const pool = getPool();
-
-        const group = await wgDbForLeader.getWorkingGroupBySlug(slug);
-        if (!group) {
-          return res.status(404).json({
-            error: 'Working group not found',
-            message: `No working group found with slug: ${slug}`,
-          });
-        }
-
-        // Check post belongs to this working group
-        const existing = await pool.query(
-          `SELECT * FROM perspectives WHERE id = $1 AND working_group_id = $2`,
-          [postId, group.id]
-        );
-
-        if (existing.rows.length === 0) {
-          return res.status(404).json({
-            error: 'Post not found',
-            message: 'Post not found in this working group',
-          });
-        }
-
-        // Validate slug if provided
-        if (post_slug) {
-          const slugPattern = /^[a-z0-9-]+$/;
-          if (!slugPattern.test(post_slug)) {
-            return res.status(400).json({
-              error: 'Invalid slug',
-              message: 'Slug must contain only lowercase letters, numbers, and hyphens',
-            });
-          }
-        }
-
-        // If status is changing to published, set published_at
-        const wasPublished = existing.rows[0].status === 'published';
-        const willBePublished = status === 'published';
-        const publishedAt = willBePublished && !wasPublished
-          ? new Date()
-          : existing.rows[0].published_at;
-
-        const result = await pool.query(
-          `UPDATE perspectives SET
-            slug = COALESCE($1, slug),
-            content_type = COALESCE($2, content_type),
-            title = COALESCE($3, title),
-            subtitle = $4,
-            category = $5,
-            excerpt = $6,
-            content = $7,
-            external_url = $8,
-            external_site_name = $9,
-            author_name = COALESCE($10, author_name),
-            author_title = $11,
-            featured_image_url = $12,
-            status = COALESCE($13, status),
-            display_order = COALESCE($14, display_order),
-            tags = $15,
-            published_at = $16,
-            is_members_only = $17,
-            updated_at = NOW()
-          WHERE id = $18 AND working_group_id = $19
-          RETURNING *`,
-          [
-            post_slug || null,
-            content_type || null,
-            title || null,
-            subtitle || null,
-            category || null,
-            excerpt || null,
-            content || null,
-            external_url || null,
-            external_site_name || null,
-            author_name || null,
-            author_title || null,
-            featured_image_url || null,
-            status || null,
-            display_order ?? null,
-            tags || null,
-            publishedAt,
-            is_members_only ?? false,
-            postId,
-            group.id,
-          ]
-        );
-
-        const updatedPost = result.rows[0];
-
-        // Send Slack notification if post was just published (status changed to published)
-        if (willBePublished && !wasPublished) {
-          notifyWorkingGroupPost({
-            workingGroupName: group.name,
-            workingGroupSlug: slug,
-            postTitle: updatedPost.title,
-            postSlug: updatedPost.slug,
-            authorName: updatedPost.author_name || 'Unknown',
-            contentType: updatedPost.content_type || 'article',
-            category: updatedPost.category || undefined,
-          }).catch(err => {
-            logger.warn({ err }, 'Failed to send Slack notification for working group post');
-          });
-        }
-
-        res.json(updatedPost);
-      } catch (error) {
-        logger.error({ err: error }, 'Update working group post error');
-        if (error instanceof Error && error.message.includes('duplicate key')) {
-          return res.status(409).json({
-            error: 'Slug already exists',
-            message: 'A post with this slug already exists',
-          });
-        }
-        res.status(500).json({
-          error: 'Failed to update post',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // DELETE /api/working-groups/:slug/manage/posts/:postId - Delete post as leader
-    this.app.delete('/api/working-groups/:slug/manage/posts/:postId', requireAuth, requireWorkingGroupLeader, async (req, res) => {
-      try {
-        const { slug, postId } = req.params;
-        const pool = getPool();
-
-        const group = await wgDbForLeader.getWorkingGroupBySlug(slug);
-        if (!group) {
-          return res.status(404).json({
-            error: 'Working group not found',
-            message: `No working group found with slug: ${slug}`,
-          });
-        }
-
-        const result = await pool.query(
-          `DELETE FROM perspectives WHERE id = $1 AND working_group_id = $2 RETURNING id`,
-          [postId, group.id]
-        );
-
-        if (result.rows.length === 0) {
-          return res.status(404).json({
-            error: 'Post not found',
-            message: 'Post not found in this working group',
-          });
-        }
-
-        res.json({ success: true, deleted: postId });
-      } catch (error) {
-        logger.error({ err: error }, 'Delete working group post error');
-        res.status(500).json({
-          error: 'Failed to delete post',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // POST /api/working-groups/:slug/manage/fetch-url - Fetch URL metadata (for link posts)
-    this.app.post('/api/working-groups/:slug/manage/fetch-url', requireAuth, requireWorkingGroupLeader, async (req, res) => {
-      try {
-        const { url } = req.body;
-
-        if (!url) {
-          return res.status(400).json({
-            error: 'URL required',
-            message: 'Please provide a URL to fetch',
-          });
-        }
-
-        // Fetch the page
-        const response = await fetch(url, {
-          headers: {
-            'User-Agent': 'Mozilla/5.0 (compatible; AgenticAdvertising/1.0)',
-            'Accept': 'text/html,application/xhtml+xml',
-          },
-          redirect: 'follow',
-        });
-
-        if (!response.ok) {
-          throw new Error(`Failed to fetch URL: ${response.status}`);
-        }
-
-        const html = await response.text();
-
-        // Extract metadata from HTML
-        const titleMatch = html.match(/<title[^>]*>([^<]+)<\/title>/i);
-        const ogTitleMatch = html.match(/<meta[^>]+property=["']og:title["'][^>]+content=["']([^"']+)["']/i)
-          || html.match(/<meta[^>]+content=["']([^"']+)["'][^>]+property=["']og:title["']/i);
-        const ogDescMatch = html.match(/<meta[^>]+property=["']og:description["'][^>]+content=["']([^"']+)["']/i)
-          || html.match(/<meta[^>]+content=["']([^"']+)["'][^>]+property=["']og:description["']/i);
-        const descMatch = html.match(/<meta[^>]+name=["']description["'][^>]+content=["']([^"']+)["']/i)
-          || html.match(/<meta[^>]+content=["']([^"']+)["'][^>]+name=["']description["']/i);
-        const ogSiteMatch = html.match(/<meta[^>]+property=["']og:site_name["'][^>]+content=["']([^"']+)["']/i)
-          || html.match(/<meta[^>]+content=["']([^"']+)["'][^>]+property=["']og:site_name["']/i);
-
-        // Helper to decode HTML entities
-        const decodeHtmlEntities = (text: string): string => {
-          return text
-            .replace(/&#x([0-9a-fA-F]+);/g, (_, hex) => String.fromCharCode(parseInt(hex, 16)))
-            .replace(/&#(\d+);/g, (_, dec) => String.fromCharCode(parseInt(dec, 10)))
-            .replace(/&quot;/g, '"')
-            .replace(/&apos;/g, "'")
-            .replace(/&lt;/g, '<')
-            .replace(/&gt;/g, '>')
-            .replace(/&amp;/g, '&');
-        };
-
-        let title = ogTitleMatch?.[1] || titleMatch?.[1] || '';
-        title = decodeHtmlEntities(title.trim());
-
-        let excerpt = ogDescMatch?.[1] || descMatch?.[1] || '';
-        excerpt = decodeHtmlEntities(excerpt.trim());
-        // Truncate excerpt to 160 chars max
-        if (excerpt.length > 160) {
-          excerpt = excerpt.substring(0, 157) + '...';
-        }
-
-        let site_name = ogSiteMatch?.[1] || '';
-        if (!site_name) {
-          try {
-            const parsedUrl = new URL(url);
-            site_name = parsedUrl.hostname.replace('www.', '');
-            site_name = site_name.charAt(0).toUpperCase() + site_name.slice(1);
-          } catch {
-            // ignore URL parse errors
-          }
-        }
-        site_name = decodeHtmlEntities(site_name);
-
-        res.json({ title, excerpt, site_name });
-      } catch (error) {
-        logger.error({ err: error }, 'Fetch URL metadata error (working group)');
-        res.status(500).json({
-          error: 'Failed to fetch URL',
-          message: error instanceof Error ? error.message : 'Unknown error',
-        });
-      }
-    });
-
-    // GET /api/working-groups/for-organization/:orgId - Get working groups that users from an org belong to
-    this.app.get('/api/working-groups/for-organization/:orgId', async (req, res) => {
-      try {
-        const { orgId } = req.params;
-        const wgDb = new WorkingGroupDatabase();
-
-        const groups = await wgDb.getWorkingGroupsForOrganization(orgId);
-        res.json({ working_groups: groups });
-      } catch (error) {
-        logger.error({ err: error }, 'Get org working groups error');
-        res.status(500).json({
-          error: 'Failed to get working groups',
           message: error instanceof Error ? error.message : 'Unknown error',
         });
       }
